Rename separate build modules
